added db test and refactored db
